+++ AuditLiteLib/Configuration/AuditConfigBuilder.cs
@@ -3,7 +3,7 @@
 public class AuditConfigBuilder
 {
     private string _predefinedConfigName = string.Empty;
-    private int _flushIntervalMilliseconds = 50000;
+    private long _flushIntervalMilliseconds = 50000;
     private int _maxBufferSize = 100;
     private string _serverUrl = "http://localhost:5000/audit";
     private bool _enableProtobufSerialization = true;
@@ -20,7 +20,7 @@ public class AuditConfigBuilder
         return this;
     }
 
-    public AuditConfigBuilder SetFlushIntervalMilliseconds(int interval)
+    public AuditConfigBuilder SetFlushIntervalMilliseconds(long interval)
     {
         _flushIntervalMilliseconds = interval;
         return this;
